refactor plan build into dedicated pipeline module (#17)
